X.26 (#627)
* Initial branch of x.26

* Error on deprecated (#618)

* Review deprecated methods, change build to error on deprecation

* <release>8</release>

* Fix compilation on Java 21

* Use release 8 only for Java 11+

* Move ByteBuffers utility to where it is used

* Fix the version range

* Initial branch of x.26

* Narrow the SuppressWarnings to the line that needs it

* Not deprecated

* Deprecate due x.26.

* Bump versions to 26ea.

* requirePositive instead of require(positive()

* fix license

* Remove se.eris.notnull.instrument

* Cast to ensure the right method is called in Java 8

* X.26 common compile flags (#620)

* Initial branch of x.26

* Error on deprecated (#618)

* Review deprecated methods, change build to error on deprecation

* <release>8</release>

* Fix compilation on Java 21

* Use release 8 only for Java 11+

* Move ByteBuffers utility to where it is used

* Fix the version range

* Initial branch of x.26

* Narrow the SuppressWarnings to the line that needs it

* Not deprecated

* Use a common configuration for the compiler-plugin where deprecation warnings cause the build to fail

* Use a common configuration for the compiler-plugin where deprecation warnings cause the build to fail

* Use a common configuration for the compiler-plugin where deprecation warnings cause the build to fail

* Use a common configuration for the compiler-plugin where deprecation warnings cause the build to fail

* Deprecate due x.26.

---------


* X.26 lint all (#621)

* review with lint:all

* review with lint:all

* Add -Xlint:all,-options as default compiler

* Tidy up compiler setting and some warnings

* Consolidate references to Thread.currentThread.getId()

* Move uncheckedCast to core/Jvm

* Add -Xlint:all,-options as default compiler

* @SuppressWarnings({"rawtypes"}) -> @SuppressWarnings("rawtypes")

* Fix deprecated methods/classes in Core

* Tidy up code base

* fpAppend0 deprecated so FP_APPEND0 can be removed

* resolve Javadoc issues

* Added fail on warning lint for Javadoc, and rawtypes fixes for Bytes (#624)

* Move tests for BytesInternal to the internal package

* Fix a check that should be max not min.

* Javadoc fix

* Error on deprecated (#618)

* Review deprecated methods, change build to error on deprecation

* <release>8</release>

* Fix compilation on Java 21

* Use release 8 only for Java 11+

* Move ByteBuffers utility to where it is used

* Fix the version range

* Initial branch of x.26

* Narrow the SuppressWarnings to the line that needs it

* Not deprecated

* X.26 lint all (#621)

* review with lint:all

* review with lint:all

* Add -Xlint:all,-options as default compiler

* Added fail on warning lint for Javadoc, and rawtypes fixes for Bytes (#624)

* Fix warnings following rebase with develop

* FIX deprecation issues in Java 17+

* Fix newlines

* Maintain backward compatibility as the return type has changed in Java 11+

* </p> tag in javadoc not allowed

* bytes -> bytesStore

* parent poms -> x.26.0

* Only deprecated in Java 17

* Where marked for removal, also mark for deprecation

---------

+++ src/test/java/net/openhft/chronicle/bytes/internal/DecimaliserFloatTest.java
@@ -75,7 +75,7 @@ class DecimaliserFloatTest extends BytesTestCommon {
             assertEquals(48847, mantissa);
             assertEquals(16, exponent);
         };
-        assertTrue(new MaximumPrecision(16).toDecimal((float) HARD_TO_DECIMALISE, check));
+        assertTrue(new MaximumPrecision(16).toDecimal(HARD_TO_DECIMALISE, check));
     }
 
     @Test
@@ -206,3 +206,4 @@ class DecimaliserFloatTest extends BytesTestCommon {
         assertTrue(UsesBigDecimal.USES_BIG_DECIMAL.toDecimal((float) Long.MIN_VALUE, check));
     }
 }
+
